+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - context: now `elfo_message_handling_time_seconds` includes pseudo messages `<Startup>` and `<EmptyMailbox>` ([#64]).
+- routers: add `Outcome::GentleUnicast` and `Outcome::GentleMulticast` ([#65]).
 
 ### Changed
 - context: reduce the budget of `recv()` from `256` to `64`.
@@ -17,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - core: `ValidateConfig`, `Terminate` and `Ping` doesn't cause spawning singletons ([#63]).
 
 [#68]: https://github.com/elfo-rs/elfo/issues/68
+[#65]: https://github.com/elfo-rs/elfo/issues/65
 [#64]: https://github.com/elfo-rs/elfo/issues/64
 [#63]: https://github.com/elfo-rs/elfo/issues/63